Refactor of ranges
The goal of the refactor was to change ownership semantics.
Before the change, ranges generally stored a pair of iterators of the source range.
This caused lifetime issues in certain situations.
Now, the source range can be moved into a new range, thus binding its lifetime
to a new range.